update deploy script
